drf_basic_CRUD/requirements.txt
asgiref==3.2.7
Django==3.0.3
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-recaptcha==2.0.6
djangorestframework==3.11.0
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
